sdk/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyflare"
version = "0.1.0"
description = "PyFlare SDK - OpenTelemetry-native observability for AI/ML workloads"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "PyFlare Team" }
]
keywords = [
    "observability",
    "monitoring",
    "ai",
    "ml",
    "machine-learning",
    "opentelemetry",
    "tracing",
    "llm",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
]

dependencies = [
    "opentelemetry-api>=1.23.0",
    "opentelemetry-sdk>=1.23.0",
    "opentelemetry-exporter-otlp>=1.23.0",
    "pydantic>=2.6.0",
    "httpx>=0.27.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "black>=24.2.0",
    "mypy>=1.8.0",
]
langchain = [
    "langchain>=0.1.0",
]
openai = [
    "openai>=1.12.0",
]
pytorch = [
    "torch>=2.0.0",
]
all = [
    "pyflare[langchain,openai,pytorch]",
]

[project.urls]
Homepage = "https://github.com/pyflare/pyflare"
Documentation = "https://pyflare.dev/docs"
Repository = "https://github.com/pyflare/pyflare"
Issues = "https://github.com/pyflare/pyflare/issues"

[tool.hatch.build.targets.wheel]
packages = ["pyflare"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
